consolidate config report cleanup fixture with conftest.py
