+++ tests/conftest.py
@@ -1,0 +1,28 @@
+import pytest
+import os
+from garak import _config
+
+
+@pytest.fixture(autouse=True)
+def config_report_cleanup(request):
+    """Cleanup a testing and report directory once we are finished."""
+
+    def remove_log_files():
+        files = []
+        if _config.transient.reportfile is not None:
+            _config.transient.reportfile.close()
+            report_html_file = _config.transient.report_filename.replace(
+                ".jsonl", ".html"
+            )
+            hitlog_file = _config.transient.report_filename.replace(
+                ".report.", ".hitlog."
+            )
+            files.append(_config.transient.report_filename)
+            files.append(report_html_file)
+            files.append(hitlog_file)
+
+        for file in files:
+            if os.path.exists(file):
+                os.remove(file)
+
+    request.addfinalizer(remove_log_files)
