+++ tests/test_config.py
@@ -485,29 +485,3 @@ def test_report_prefix_with_hitlog_no_explode():
     assert os.path.isfile("kjsfhgkjahpsfdg.report.jsonl")
     assert os.path.isfile("kjsfhgkjahpsfdg.report.html")
     assert os.path.isfile("kjsfhgkjahpsfdg.hitlog.jsonl")
-
-
-@pytest.fixture(autouse=True)
-def cleanup(request):
-    """Cleanup a testing and report directory once we are finished."""
-
-    def remove_log_files():
-        files = [
-            "laurelhurst.report.jsonl",
-            "kjsfhgkjahpsfdg.report.jsonl",
-            "kjsfhgkjahpsfdg.report.html",
-            "kjsfhgkjahpsfdg.hitlog.jsonl",
-        ]
-        if _config.transient.reportfile is not None:
-            _config.transient.reportfile.close()
-            report_html_file = _config.transient.report_filename.replace(
-                ".jsonl", ".html"
-            )
-            files.append(_config.transient.report_filename)
-            files.append(report_html_file)
-
-        for file in files:
-            if os.path.exists(file):
-                os.remove(file)
-
-    request.addfinalizer(remove_log_files)
